Extract PostQuery/convert to js, try to get clear working
