--- webapp/graphql/PostQuery.gql
@@ -1,66 +1,0 @@
-query Post($slug: String!) {
-  Post(slug: $slug) {
-    id
-    title
-    content
-    createdAt
-    disabled
-    deleted
-    slug
-    image
-    author {
-      id
-      slug
-      name
-      avatar
-      disabled
-      deleted
-      shoutedCount
-      contributionsCount
-      commentsCount
-      followedByCount
-      followedByCurrentUser
-      badges {
-        id
-        key
-        icon
-      }
-    }
-    tags {
-      name
-    }
-    commentsCount
-    comments(orderBy: createdAt_desc) {
-      id
-      contentExcerpt
-      createdAt
-      disabled
-      deleted
-      author {
-        id
-        slug
-        name
-        avatar
-        disabled
-        deleted
-        shoutedCount
-        contributionsCount
-        commentsCount
-        followedByCount
-        followedByCurrentUser
-        badges {
-          id
-          key
-          icon
-        }
-      }
-    }
-    categories {
-      id
-      name
-      icon
-    }
-    shoutedCount
-    shoutedByCurrentUser
-  }
-}
